:sparkles: BE/Feat [S08P31A702-36] : Chatting add API
Chatting 관련 DB및 Table 변경 예정

+++ backend-business/FEELINGFILLING/src/main/java/com/a702/feelingfilling/domain/chatting/model/dto/ChatInputDTO.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ChatInputDTO {
   String content;
-  short type;
+  int type;
 }
